backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==2.1.1
geopandas==1.0.1
fiona==1.10.1
shapely==2.0.6
svgwrite==1.4.3
scipy==1.14.1
python-multipart==0.0.12
requests==2.32.3
